Fix the chaining award calculation
Previously the chaining award could be broken if two people in the
middle of the chain had the same time. E.g. 21, 22, 22, 23 would not
unlock the Chain 3 award.

I've updated it so that the award is unlocked in that scenario. A
chain's length is determined by the number of sequential times, rather
than the number of people involved. So times of 21, 21, 22, 22, 23, 23
will unlock Chain 3 rather than Chain 6.