Handle glitched times
Times can be marked as glitched, and doing this means that they no
longer count as a person's best time or contribute to that person's
average time.

Glitched times do still count for streak and award calculation. This is
because editing times after the fact completely adjusts the leaderboards
and results in lots of award changes.