Commit some horrible horrible crimes
So New York Times updated their leaderboard, I was afraid this would
happen one day but assumed I'd just be able to change my class names and
continue scraping the HTML like a normal person.

Alas.

The New York Times have decided that displaying a simple table with
names and times should absolutely 100% rely on client-side JavaScript 👍

I'm not ready to let this leaderboard go and so I've done something
horrible. The leaderboard page is completely blank if you load it
without JavaScript _however_ weirdly the leaderboard data is still in
the DOM just as an object inside a `<script>` element. I decided the
best thing to do would be to execute this untrusted JavaScript and see
if I can get the data out of it.

Turns out I can but not without taking at least 4d6 psychic damage.